refactoring pooling strategy. now only used for graphing, not alerts. -- fastnetmon calls API for pool/alerting now
